refactored product deleting

+++ src/main/java/kg/kstu/production/service/impl/ProductServiceImpl.java
@@ -1,11 +1,15 @@
 package kg.kstu.production.service.impl;
 
+import jakarta.transaction.Transactional;
+import kg.kstu.production.entity.Ingredient;
 import kg.kstu.production.entity.Product;
+import kg.kstu.production.repository.IngredientsRepository;
 import kg.kstu.production.repository.ProductRepository;
 import kg.kstu.production.service.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.beans.Transient;
 import java.util.List;
 import java.util.Optional;
 
@@ -13,6 +17,8 @@ import java.util.Optional;
 public class ProductServiceImpl implements ProductService {
     @Autowired
     private ProductRepository productRepository;
+    @Autowired
+    private IngredientsRepository ingredientsRepository;
 
     @Override
     public List<Product> getAll() {
@@ -25,8 +31,16 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public void deleteProduct(Long id) {
+    @Transactional
+    public String deleteProduct(Long id) {
+        Optional<Product> product = productRepository.findById(id);
+        if(product.isPresent()) {
+        ingredientsRepository.deleteAllByProduct(product);
         productRepository.deleteById(id);
+        return "Done";
+        } else {
+            return "Error";
+        }
     }
 
     @Override
